test: updating subaccount tests
